Removed internal/CoreCommands and moved the commands into commands/CoreCommands
